Extract gapped page cursors class
