+++ lib/gapped_page_cursors.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class BasicPageCursors
+class GappedPageCursors
   def self.as_hash(total_pages, current_page, per_page)
     new(total_pages, current_page, per_page).to_hash
   end
@@ -16,6 +16,8 @@ class BasicPageCursors
   def to_hash
     {
       around: around_cursors,
+      first: first_cursor,
+      last: last_cursor,
       previous: previous_cursor
     }.compact
   end
@@ -24,6 +26,20 @@ class BasicPageCursors
 
   def around_cursors
     around_page_numbers.map { |page_num| page_cursor(page_num) }
+  end
+
+  def first_cursor
+    include_first_cursor = !around_page_numbers.include?(1)
+    return unless include_first_cursor
+
+    page_cursor(1)
+  end
+
+  def last_cursor
+    include_last_cursor = !around_page_numbers.include?(total_pages)
+    return unless include_last_cursor
+
+    page_cursor(total_pages)
   end
 
   def previous_cursor
@@ -38,6 +54,12 @@ class BasicPageCursors
   end
 
   def around_page_numbers
-    (1..total_pages).to_a
+    if current_page <= 3
+      (1..4).to_a
+    elsif current_page >= total_pages - 2
+      ((total_pages - 3)..total_pages).to_a
+    else
+      [current_page - 1, current_page, current_page + 1]
+    end
   end
 end
