+++ lib/page_cursor_resolver.rb
@@ -2,6 +2,7 @@
 
 require 'page_cursor'
 require 'basic_page_cursors'
+require 'gapped_page_cursors'
 
 class PageCursorResolver
   MAX_CURSOR_COUNT = 5
@@ -13,18 +14,13 @@ class PageCursorResolver
     @context = context
   end
 
-  def page_cursors # rubocop:disable Metrics/MethodLength
+  def page_cursors
     return if total_pages <= 1
 
     if total_pages <= MAX_CURSOR_COUNT
       BasicPageCursors.as_hash(total_pages, current_page, per_page)
     else
-      {
-        around: around_cursors,
-        first: first_cursor,
-        last: last_cursor,
-        previous: previous_cursor
-      }.compact
+      GappedPageCursors.as_hash(total_pages, current_page, per_page)
     end
   end
 
@@ -44,51 +40,6 @@ class PageCursorResolver
     return object.nodes unless object.respond_to?(:items)
 
     object.items
-  end
-
-  def around_cursors
-    around_page_numbers.map { |page_num| page_cursor(page_num) }
-  end
-
-  def first_cursor
-    exceeds_max_cursor_count = total_pages > MAX_CURSOR_COUNT
-    include_first_cursor =
-      exceeds_max_cursor_count && !around_page_numbers.include?(1)
-    return unless include_first_cursor
-
-    page_cursor(1)
-  end
-
-  def last_cursor
-    exceeds_max_cursor_count = total_pages > MAX_CURSOR_COUNT
-    include_last_cursor =
-      exceeds_max_cursor_count && !around_page_numbers.include?(total_pages)
-    return unless include_last_cursor
-
-    page_cursor(total_pages)
-  end
-
-  def previous_cursor
-    include_previous_cursor = current_page > 1
-    return unless include_previous_cursor
-
-    page_cursor(current_page - 1)
-  end
-
-  def page_cursor(page_number)
-    PageCursor.as_hash(current_page, page_number, per_page)
-  end
-
-  def around_page_numbers # rubocop:disable Metrics/AbcSize
-    if total_pages <= MAX_CURSOR_COUNT
-      (1..total_pages).to_a
-    elsif current_page <= 3
-      (1..4).to_a
-    elsif current_page >= total_pages - 2
-      ((total_pages - 3)..total_pages).to_a
-    else
-      [current_page - 1, current_page, current_page + 1]
-    end
   end
 
   def current_page
